requirements.txt
connexion>=1.0.31
gevent>=1.2.2
PyYAML>=3.10
jsonschema>=2.6.0
